fix 'long' type bugs

+++ ds/hw1/p3.cc
@@ -43,7 +43,7 @@ void right_j(long A[], int n, long long m, int out[]) {
     out[n-1] = n-1;
 }
 
-int solve(int A[], int n, int m) {
+int solve(long A[], int n, int m) {
     int right[n];
     right_j(A, n, m, right);
     // print(right, n);
@@ -58,21 +58,21 @@ int solve(int A[], int n, int m) {
     tests
 */
 
-int* ex1(int &n, int &m) {
+long* ex1(int &n, int &m) {
     n = 6;
     m = 6;
-    return new int[n] {6, 5, 2, 2, 6, 6};
+    return new long[n] {6, 5, 2, 2, 6, 6};
 }
 
-int* ex2(int &n, int &m) {
+long* ex2(int &n, int &m) {
     n = 5;
     m = 3;
-    return new int[n] {7, 6, 5, 4, 3};
+    return new long[n] {7, 6, 5, 4, 3};
 }
 
 int test() {
     int n, m;
-    int *A = new int[n];
+    long *A = new long[n];
     A = ex2(n,m);
     std::cout << solve(A, n, m);
     return 0;
